+++ spec/front_end/parser_spec.rb
@@ -37,7 +37,7 @@ module Loxxy
           # Parse results MUST to comply to grammar rule:
           #   program => declaration_star EOF
           #   where the declaration_star MUST be empty
-          expect(aParseTree.root.symbol.name).to eq('EOF')
+          expect(aParseTree.root).to be_kind_of(Ast::LoxNoopExpr)
         end
 
         it 'should cope with an empty input' do
@@ -64,7 +64,7 @@ module Loxxy
         it 'should parse a false literal' do
           input = 'false;'
           ptree = subject.parse(input)
-          leaf = walk_subnodes(ptree.root, [0, 0])
+          leaf = ptree.root.subnodes[0]
           expect(leaf).to be_kind_of(Ast::LoxLiteralExpr)
           expect(leaf.literal).to be_equal(Datatype::False.instance)
         end
@@ -72,7 +72,7 @@ module Loxxy
         it 'should parse a true literal' do
           input = 'true;'
           ptree = subject.parse(input)
-          leaf = walk_subnodes(ptree.root, [0, 0])
+          leaf = ptree.root.subnodes[0]
           expect(leaf).to be_kind_of(Ast::LoxLiteralExpr)
           expect(leaf.literal).to be_equal(Datatype::True.instance)
         end
@@ -81,7 +81,7 @@ module Loxxy
           inputs = %w[1234; 12.34;]
           inputs.each do |source|
             ptree = subject.parse(source)
-            leaf = walk_subnodes(ptree.root, [0, 0])
+            leaf = ptree.root.subnodes[0]
             expect(leaf).to be_kind_of(Ast::LoxLiteralExpr)
             expect(leaf.literal).to be_kind_of(Datatype::Number)
             expect(leaf.literal.value).to eq(source.to_f)
@@ -96,7 +96,7 @@ module Loxxy
           ]
           inputs.each do |source|
             ptree = subject.parse(source)
-            leaf = walk_subnodes(ptree.root, [0, 0])
+            leaf = ptree.root.subnodes[0]
             expect(leaf).to be_kind_of(Ast::LoxLiteralExpr)
             expect(leaf.literal).to be_kind_of(Datatype::LXString)
             expect(leaf.literal.value).to eq(source.gsub(/(^")|(";$)/, ''))
@@ -106,7 +106,7 @@ module Loxxy
         it 'should parse a nil literal' do
           input = 'nil;'
           ptree = subject.parse(input)
-          leaf = walk_subnodes(ptree.root, [0, 0])
+          leaf = ptree.root.subnodes[0]
           expect(leaf).to be_kind_of(Ast::LoxLiteralExpr)
           expect(leaf.literal).to be_equal(Datatype::Nil.instance)
         end
@@ -119,21 +119,11 @@ module Loxxy
             print "Hello, world!";
 LOX_END
           ptree = subject.parse(program)
-          root = ptree.root
-          expect(root.symbol.name).to eq('program')
-          (prnt_stmt, eof) = root.subnodes
-          expect(prnt_stmt).to be_kind_of(Rley::PTree::NonTerminalNode)
-          expect(prnt_stmt.symbol.name).to eq('printStmt')
-          expect(prnt_stmt.subnodes.size).to eq(3)
-          expect(prnt_stmt.subnodes[0]).to be_kind_of(Rley::PTree::TerminalNode)
-          expect(prnt_stmt.subnodes[0].symbol.name).to eq('PRINT')
-          expect(prnt_stmt.subnodes[1]).to be_kind_of(Loxxy::Ast::LoxLiteralExpr)
-          expect(prnt_stmt.subnodes[1].literal).to be_kind_of(Loxxy::Datatype::LXString)
-          expect(prnt_stmt.subnodes[1].literal.value).to eq('Hello, world!')
-          expect(prnt_stmt.subnodes[2]).to be_kind_of(Rley::PTree::TerminalNode)
-          expect(prnt_stmt.subnodes[2].symbol.name).to eq('SEMICOLON')
-          expect(eof).to be_kind_of(Rley::PTree::TerminalNode)
-          expect(eof.symbol.name).to eq('EOF')
+          prnt_stmt = ptree.root
+          expect(prnt_stmt).to be_kind_of(Ast::LoxPrintStmt)
+          expect(prnt_stmt.subnodes[0]).to be_kind_of(Ast::LoxLiteralExpr)
+          expect(prnt_stmt.subnodes[0].literal).to be_kind_of(Loxxy::Datatype::LXString)
+          expect(prnt_stmt.subnodes[0].literal.value).to eq('Hello, world!')
         end
       end # context
 
@@ -141,7 +131,7 @@ LOX_END
         it 'should parse the addition of two number literals' do
           input = '123 + 456;'
           ptree = subject.parse(input)
-          parent = ptree.root.subnodes[0]
+          parent = ptree.root
           expect(parent).to be_kind_of(Rley::PTree::NonTerminalNode)
           expect(parent.symbol.name).to eq('exprStmt')
           expr = parent.subnodes[0]
@@ -154,7 +144,7 @@ LOX_END
         it 'should parse the subtraction of two number literals' do
           input = '4 - 3;'
           ptree = subject.parse(input)
-          parent = ptree.root.subnodes[0]
+          parent = ptree.root
           expect(parent).to be_kind_of(Rley::PTree::NonTerminalNode)
           expect(parent.symbol.name).to eq('exprStmt')
           expr = parent.subnodes[0]
@@ -167,7 +157,7 @@ LOX_END
         it 'should parse multiple additive operations' do
           input = '5 + 2 - 3;'
           ptree = subject.parse(input)
-          parent = ptree.root.subnodes[0]
+          parent = ptree.root
           expect(parent).to be_kind_of(Rley::PTree::NonTerminalNode)
           expect(parent.symbol.name).to eq('exprStmt')
           expr = parent.subnodes[0]
@@ -183,7 +173,7 @@ LOX_END
         it 'should parse the division of two number literals' do
           input = '8 / 2;'
           ptree = subject.parse(input)
-          parent = ptree.root.subnodes[0]
+          parent = ptree.root
           expect(parent).to be_kind_of(Rley::PTree::NonTerminalNode)
           expect(parent.symbol.name).to eq('exprStmt')
           expr = parent.subnodes[0]
@@ -196,7 +186,7 @@ LOX_END
         it 'should parse the product of two number literals' do
           input = '12.34 * 0.3;'
           ptree = subject.parse(input)
-          parent = ptree.root.subnodes[0]
+          parent = ptree.root
           expect(parent).to be_kind_of(Rley::PTree::NonTerminalNode)
           expect(parent.symbol.name).to eq('exprStmt')
           expr = parent.subnodes[0]
@@ -209,7 +199,7 @@ LOX_END
         it 'should parse multiple multiplicative operations' do
           input = '5 * 2 / 3;'
           ptree = subject.parse(input)
-          parent = ptree.root.subnodes[0]
+          parent = ptree.root
           expect(parent).to be_kind_of(Rley::PTree::NonTerminalNode)
           expect(parent.symbol.name).to eq('exprStmt')
           expr = parent.subnodes[0]
@@ -225,7 +215,7 @@ LOX_END
         it 'should parse combination of terms and factors' do
           input = '5 + 2 / 3;'
           ptree = subject.parse(input)
-          parent = ptree.root.subnodes[0]
+          parent = ptree.root
           expect(parent).to be_kind_of(Rley::PTree::NonTerminalNode)
           expect(parent.symbol.name).to eq('exprStmt')
           expr = parent.subnodes[0]
@@ -243,7 +233,7 @@ LOX_END
         it 'should parse the concatenation of two string literals' do
           input = '"Lo" + "ve";'
           ptree = subject.parse(input)
-          parent = ptree.root.subnodes[0]
+          parent = ptree.root
           expect(parent).to be_kind_of(Rley::PTree::NonTerminalNode)
           expect(parent.symbol.name).to eq('exprStmt')
           expr = parent.subnodes[0]
@@ -259,7 +249,7 @@ LOX_END
         %w[> >= < <=].each do |predicate|
           input = "3 #{predicate} 2;"
           ptree = subject.parse(input)
-          parent = ptree.root.subnodes[0]
+          parent = ptree.root
           expect(parent).to be_kind_of(Rley::PTree::NonTerminalNode)
           expect(parent.symbol.name).to eq('exprStmt')
           expr = parent.subnodes[0]
@@ -276,7 +266,7 @@ LOX_END
           %w[!= ==].each do |predicate|
             input = "3 #{predicate} 2;"
             ptree = subject.parse(input)
-            parent = ptree.root.subnodes[0]
+            parent = ptree.root
             expect(parent).to be_kind_of(Rley::PTree::NonTerminalNode)
             expect(parent.symbol.name).to eq('exprStmt')
             expr = parent.subnodes[0]
@@ -290,7 +280,7 @@ LOX_END
         it 'should parse combination of equality expressions' do
           input = '5 != 2 == false; // A bit contrived example'
           ptree = subject.parse(input)
-          parent = ptree.root.subnodes[0]
+          parent = ptree.root
           expect(parent).to be_kind_of(Rley::PTree::NonTerminalNode)
           expect(parent.symbol.name).to eq('exprStmt')
           expr = parent.subnodes[0]
@@ -309,7 +299,7 @@ LOX_END
           %w[or and].each do |connector|
             input = "5 > 2 #{connector} 3 <= 4;"
             ptree = subject.parse(input)
-            parent = ptree.root.subnodes[0]
+            parent = ptree.root
             expect(parent).to be_kind_of(Rley::PTree::NonTerminalNode)
             expect(parent.symbol.name).to eq('exprStmt')
             expr = parent.subnodes[0]
@@ -329,7 +319,7 @@ LOX_END
         it 'should parse a combinations of logical expressions' do
           input = '4 > 3 and 1 < 2 or 4 >= 5;'
           ptree = subject.parse(input)
-          parent = ptree.root.subnodes[0]
+          parent = ptree.root
           expect(parent).to be_kind_of(Rley::PTree::NonTerminalNode)
           expect(parent.symbol.name).to eq('exprStmt')
           expr = parent.subnodes[0]